move other tests shared functions to /tests/utils/
requireUncached will be used in other PRs, for example to test
config.js in #320
