+++ test/utils/requireUncached.js
@@ -1,0 +1,6 @@
+function requireUncached(module) {
+    delete require.cache[require.resolve(module)];
+    return require(module);
+}
+
+exports.requireUncached = requireUncached;
